getMidnightDate() moved from CalendarEngineHQL to Calendar
